+++ tests/integration_tests.rs
@@ -78,8 +78,8 @@ fn it_works() {
                 ..Style::default()
             },
         })
-        .with_quote_highlighter(QuotesConfig {
-            quotes_token: b'"',
+        .with_quote_highlighter(QuoteConfig {
+            quote_token: b'"',
             style: Style {
                 fg: Some(Color::Yellow),
                 ..Style::default()